move the slicer
